Move resource types into gl/types.hpp
