+++ lib/gl/renderer.cpp
@@ -115,15 +115,15 @@ void MeshRenderer::render(const RenderQueue& queue)
    {
       return !renderPair.first.contains_translucent;
    });
-   GLDevice::RenderBufHandle renderBufs[2];
-   GLDevice::FBOHandle msaaFb;
+   RenderBufHandle renderBufs[2];
+   FBOHandle msaaFb;
    if (feat_use_fbo_antialias && msaa_enable)
    {
       GLuint colorBuf, depthBuf;
       glGenRenderbuffers(1, &colorBuf);
       glGenRenderbuffers(1, &depthBuf);
-      renderBufs[0] = GLDevice::RenderBufHandle(colorBuf);
-      renderBufs[1] = GLDevice::RenderBufHandle(depthBuf);
+      renderBufs[0] = RenderBufHandle(colorBuf);
+      renderBufs[1] = RenderBufHandle(depthBuf);
 
       GLuint fbo;
       glGenFramebuffers(1, &fbo);
@@ -154,7 +154,7 @@ void MeshRenderer::render(const RenderQueue& queue)
       }
       else
       {
-         msaaFb = GLDevice::FBOHandle(fbo);
+         msaaFb = FBOHandle(fbo);
       }
 #ifndef __EMSCRIPTEN__
       glEnable(GL_MULTISAMPLE);
@@ -252,11 +252,11 @@ void MeshRenderer::render(const RenderQueue& queue)
       int height = vp[3];
       GLuint colorBufId;
       glGenRenderbuffers(1, &colorBufId);
-      GLDevice::RenderBufHandle colorBuf(colorBufId);
+      RenderBufHandle colorBuf(colorBufId);
 
       GLuint fboId;
       glGenFramebuffers(1, &fboId);
-      GLDevice::FBOHandle resolveFb(fboId);
+      FBOHandle resolveFb(fboId);
 
       glBindRenderbuffer(GL_RENDERBUFFER, colorBuf);
       glRenderbufferStorage(GL_RENDERBUFFER, GL_RGBA8, width, height);
